requirements.txt
streamlit
PyMuPDF
requests
Pillow